requirements.txt
fastapi
uvicorn
llama-index
openai
python-dotenv
